DarkEdif: Fix C++20 building
Some brackets were needed to prevent VS2022 confusing Edif::SDK type with Edif:::SDK singleton.
Some extra string casts were needed for json_value as well - maybe std::string constructing from const char * was made explicit in C++20?

Also some cleanup to remove code warnings.

+++ DarkEdif/Inc/Android/MMFAndroidMasterHeader.h
@@ -276,7 +276,7 @@ struct qualToOi {
 
 struct RunHeader {
 	NO_DEFAULT_CTORS(RunHeader);
-	unsigned short GetEventCount();
+	int GetEventCount();
 };
 
 struct CreateObjectInfo {
